% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_region.R
\name{add_region}
\alias{add_region}
\title{Add the regions to a plot}
\usage{
add_region(obs_data)
}
\arguments{
\item{obs_data}{a dataframe of observations}
}
\value{
the same dataframe with one new column, NOM_PROV_N. This new column contains the French names of the natural provinces of Quebec
}
\description{
This function adds the CERQ province names to any dataframe.
}
\details{
Observations from anywhere in Quebec must fall in one of the 20 CERQ regions.
In most of our apps, we use these regions to select subsets of the data. 
To do this, we need to know the region which matches each observation. 
This function automates that process. 

Note that the mapping may not be perfect, tk link to documentation.

This function uses sf, and will not work unless that is installed.

IMPORTANT: note that you need to have columns called specifically "longitude" and "latitude" in your data frame
}
