% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_site_name.R
\name{add_site_name_df}
\alias{add_site_name_df}
\title{Make the site name}
\usage{
add_site_name_df(cell_data)
}
\arguments{
\item{cell_data}{a data.frame containing cell info

  This data.frame must have columns "type" and "cell.name". They will be
  stuck together and simply formatted.}
}
\value{

}
\description{
Make the site name
}
