% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_empty_campaigns.R
\name{drop_empty_campaigns}
\alias{drop_empty_campaigns}
\title{drop any empty campaigns}
\usage{
drop_empty_campaigns(site_df)
}
\arguments{
\item{site_df}{data.frame of site information, as returned by rcoleo::get_sites}
}
\value{
data.frame with no empty campaigns
}
\description{
drop any empty campaigns
}
