% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ui.R
\name{fa_dependency}
\alias{fa_dependency}
\title{Depend on fontawesome css for the map icons}
\usage{
fa_dependency()
}
\value{
htmlDependency on font-awesome
}
\description{
Any figure that intends to use font awesome classes for the icons needs this
somewhere. this is included by default in tableau_de_bord(), but it is also
available for use in any fluidPage, tagList, etc.
}
