% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_rcoleo_site.R
\name{get_rcoleo_sites_sf}
\alias{get_rcoleo_sites_sf}
\title{Get rcoleo sites as sf dataframe}
\usage{
get_rcoleo_sites_sf(...)
}
\arguments{
\item{datasets}{}
}
\description{
download all the sites and get their types as well
}
