% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_taxo.R
\name{get_species_photo}
\alias{get_species_photo}
\title{Search for species photo and return url, attribution}
\usage{
get_species_photo(name)
}
\arguments{
\item{name}{name to search by}
}
\value{
the photo url, thumbnail url and comment for the photo
}
\description{
Search for species photo and return url, attribution
}
