% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_rcoleo_site.R
\name{make_icon_adders}
\alias{make_icon_adders}
\title{Create the icons for a map}
\usage{
make_icon_adders(
  site_info_sf,
  site_icons = make_site_icons(),
  habitat_col = "type",
  site_id_col = "site_code"
)
}
\arguments{
\item{site_info_sf}{dataset with site information. should be of type sf and have a column points in sf format}

\item{habitat_col}{column name that holds the habitat type. defaults to "type"}

\item{site_id_col}{column name that holds the value that gets returned when clicked. defaults to "site_code}
}
\description{
Take a site dataframe, select the icons that should go on it, and configure
these with options. The most important is to define what value should be
returned by a click.
}
