% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_map_maker.R
\name{make_leaflet_map}
\alias{make_leaflet_map}
\title{Plot a leaflet map}
\usage{
make_leaflet_map(
  mapdata = mapselector::CERQ,
  label = TRUE,
  region_name = "NOM_PROV_N"
)
}
\arguments{
\item{mapdata}{the data that supplies the things that go on your map. default is `mapselector::CERQ`}

\item{label}{should a mouseover label appear? default is TRUE}

\item{region_name}{quoted name of the column in your dataset that gives the mouseover text. Will also be returned when you click}
}
\description{
function for making a leaflet map of the chosen region
}
