% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_site_name.R
\name{make_lookup_vector}
\alias{make_lookup_vector}
\title{Create a lookup table from a data.frame}
\usage{
make_lookup_vector(some_df, value_col, name_col)
}
\arguments{
\item{some_df}{any data frame, containing two columns: one for values, one
for names used to choose a value}

\item{value_col}{the values to be looked up}

\item{name_col}{the names to use to find these values.}
}
\value{
a named vector of
}
\description{
Sometimes one column of a data.frame has human-readable names, and the other
had computer-useful codes. We often need to look up one based on the other.
This function creates a handy lookup vector from two such columns of a
data.frame. The vector is only as long at the number of distinct name-value
pairs.
}
