% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_site_name.R
\name{make_site_name}
\alias{make_site_name}
\title{take a clicked name and lookup the matching value from a vector}
\usage{
make_site_name(got_clicked_site_val, cell_lookup_vec)
}
\arguments{
\item{got_clicked_site_val}{clicked value}

\item{cell_lookup_vec}{lookup vector}
}
\value{

}
\description{
The name of the clicked site needs to be among the names in the lookup vector
}
