% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ui.R
\name{marcel}
\alias{marcel}
\title{Generate a div for the sliding explainer on the side}
\usage{
marcel(filename)
}
\arguments{
\item{filename}{the filename of the file, but not the path.}
}
\description{
Write and store the text you want in `inst/app/www/` . 
It has to be in Markdown format
}
