% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_modal_make.R
\name{mod_modal_helpbutton_ui}
\alias{mod_modal_helpbutton_ui}
\title{make a modal help button}
\usage{
mod_modal_helpbutton_ui(id, text_before_button)
}
\arguments{
\item{id}{id for the module. match it to the module that makes the modal}

\item{text_before_button}{the text that should be followed by the "i" for information.}
}
\value{

}
\description{
this function takes some text and makes a little help button appear next to it.
}
