% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_modal_contents.R
\name{modal_tab_format}
\alias{modal_tab_format}
\title{Format modal tabs}
\usage{
modal_tab_format(region, title_fmt, ...)
}
\description{
This function is a wrapper around modal tabs. It does two things. first, it
sets the title for the whole modal. Its assumed that this is going to be
based on the id of the thing you clicked on the map. Second, it formats the
modals' CSS: it applies nice CSS classes to all tabs and wraps them in modalDialog.
}
