% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_rcoleo_site.R
\name{plot_markers_controls}
\alias{plot_markers_controls}
\title{Add markers to a map of sites}
\usage{
plot_markers_controls(site_info_sf, markers)
}
\arguments{
\item{site_info_sf}{the site info df. should be sf}

\item{markers}{a list of markers to add}
}
\description{
Take a blank map and put markers on it. Markers are supplied in a list so
that they can be grouped and then controlled together.
}
