% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_type_campaign.R
\name{subset_site_df}
\alias{subset_site_df}
\title{subset the dataframe of sites to sites with specific observations}
\usage{
subset_site_df(downloaded_sites, campaign_type = NULL)
}
\arguments{
\item{downloaded_sites}{data frame of sites}

\item{campaign_type}{type of observation. can be a vector. must be one of "végétation", "papilionidés", "acoustique", "insectes_sol", 
"mammifères", "odonates", "zooplancton"}
}
\value{

}
\description{
subset the dataframe of sites to sites with specific observations
}
