% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_type_campaign.R
\name{subset_type_campaign}
\alias{subset_type_campaign}
\title{subset a list of campaigns to have all of a certain type}
\usage{
subset_type_campaign(campaign_list, campaign_type)
}
\arguments{
\item{campaign_list}{the list of dataframes containing all campaign information. usually a list column}

\item{campaign_type}{the types to filter for.}
}
\value{

}
\description{
The return from get_sites has all the campaigns listed. This function subsets that list-column to 
create a new one, containing only campaigns of a specific type.
}
