% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ui.R
\name{tab_map}
\alias{tab_map}
\title{Create a leaflet map}
\usage{
tab_map(
  title = "Map",
  id = "map",
  outputFunction = leaflet::leafletOutput,
  on_map_widgets = list(NULL)
)
}
\description{
This map creates a leaflet map. It takes an optional list of widgets to display on the map and 
places them on top.
}
